update (diff-file task) : helpers
