+++ app/helpers.php
@@ -2,16 +2,42 @@
 
 use Core\Http\Request;
 
-function checkUser() {
-     $request = new Request;
-     $user = $request->getUser();
-     if ($user !== null && $user['role_id'] == 1) {
-          return true;
-     } else {
-          return false;
-     } 
+if (! function_exists('checkUser')) {
+     function checkUser() {
+          $request = new Request;
+          $user = $request->getUser();
+          if ($user !== null && $user['role_id'] == 1) {
+               return true;
+          } else {
+               return false;
+          } 
+     }
 }
 
-function test(){
-    return 1;
+if (! function_exists('setDefineArray')) {
+     function setDefineArray($name, $ary) {
+          if ($name == "") return;
+          global $$name;
+          if (isset($$name)) return;
+          $temp = array();
+          foreach ($ary as $key => $value) {
+               $temp[$key] = $value;
+          }
+          $$name = $temp;
+          return $$name;
+     }
+}
+
+if (! function_exists('setTempGlobal')) {
+     function setTempGlobal($variableGLOBALS, $globalsVarName, $tempGlobal){
+          if (!empty($variableGLOBALS)) {
+               foreach($variableGLOBALS as $each){
+                    $globalsVarName[] = $each;
+                    $tempGlobal[$each] = $GLOBALS[$each];
+                    unset($GLOBALS[$each]);
+               }
+          }
+     
+     return array( $globalsVarName, $tempGlobal );
+     }
 }
